+++ User/util.h
@@ -26,7 +26,7 @@ insertion_sort(T* to_sort, const size_t& to_sort_size, F cmp_func)
 	for (size_t i = 1; i < to_sort_size; i++)
 	{
 		T key = to_sort[i];
-		int j = i - 1;
+		int j = (int)i - 1;
 		while (j >= 0 && !cmp_func(to_sort[j], key))
 		{
 			to_sort[j + 1] = to_sort[j];
